feat: add final onchain-summer carousel
